urf-killswitch: Add pivot killswitches for every type
A pivot killswitch represents the state of its own type and basically
is generated by a platform driver, since the platform driver actually
controls the killswitches of the system. If there is no such killswitch,
the killswitch with the smallest number will be chosen as the pivot
killswitch for its own type.