urf-killswitch: Adopt force_sync to replace key_control
Since I only encounter the killswitches sync problem in Thinkpad
X200, it seems not necessary to sync the killswitches by urfkilld
for all machines. In fact, it conflicts with the platform driver,
msi-laptop, since msi-laptop has to wait 0.5 sec before updating
the states of killswitches, and urfkilld would be confused and sync
the wrong states.