urf-config: Scan the profile dir to parse all profiles
Previously, urfkilld only matched the rules in ${vendor}-profile.xml.
However, the names of some vendors may not be suitable for this kind
of search. For example, the vendor name of EEEPC 1005HA is
"ASUSTeK Computer INC." which is long and ugly for a file name.
Scanning the whole dir could be bad for the first time startup of
urfkilld, but urfkilld eventually creates the corresponding config
file, so be it.