New GObjects UrfConsolekit and UrfSeat
UrfConsolekit contacts with ConsoleKit and maintains
the lists of "seats" and "inhibitors". UrfSeat objects
map to the seats and monitor the changes of sessions.

When the client of urfkilld inhibits its own session,
UrfDaemon registers the session id to UrfConsolekit.

When the active session is changed or a session is
removed, UrfSeat emits a signal to UrfConsolekit, and
then UrfConsolekit adjusts the state of "inhibit"
according to the list of inhibitors.