urf-device: Get the name and subsystem by libudev
Since libudev provides various functions to query the attributes
of the killswitch device, we can get the name and the subsystem
by libudev. The subsystem is especially useful when urfkilld
judges whehter the device is generated by a platform driver or
not. Now we can abandon the unreliable name matching method.